cleanup: 🧹 Jules05: Technical debt cleanup — architectural harmonization
- Consolidate RiskEngine and RiskManager into a unified 8-layer RiskManager.
- Relocate FeatureEngineer to src/data/feature_engineering.py for domain separation.
- Update main execution loop to use the harmonized validate_signal() API.
- Standardize RiskDecision return type and ATR-based position sizing.
- Document technical debt in docs/technical-debt/DEBT_LOG.md.
- Update imports system-wide (main.py, src/core, src/trading, and tests).
- Verified with 39 passing tests.
