+++ docs/technical-debt/DEBT_LOG.md
@@ -1,75 +1,43 @@
 # Technical Debt Log
 
-This log tracks architectural drift, code quality degradation, and fragmented logic introduced during multi-agent development.
+This log tracks architectural drift, code quality degradation, and fragmentation introduced by multi-agent parallelism.
 
 ## Active Debt Items
 
-### Debt Item: Legacy Temporal Markers
-**Category:** Quality
+### Debt Item: Risk Management Fragmentation
+**Category:** Fragmentation | Duplication
+**Impact:** High
+**Effort:** M
+**Resolution plan:** Harmonize `RiskManager` and `RiskEngine`. Port 8-layer ATR-based logic to `RiskManager`, standardize on `validate_signal()` API, and deprecate `RiskEngine`.
+**Owner:** Jules05 (Immediate Cleanup)
+
+### Debt Item: Feature Engineering Misplacement
+**Category:** Naming | Fragmentation
 **Impact:** Medium
 **Effort:** S
-**Resolution plan:** Replace all `datetime.utcnow()` calls with `datetime.now(timezone.utc)` for Python 3.12 compatibility and standardization.
-**Owner:** Jules05 (Immediate cleanup)
+**Resolution plan:** Relocate `src/core/feature_engineering.py` to `src/data/feature_engineering.py` to align with modular domain separation standards. Update all system-wide imports.
+**Owner:** Jules05 (Immediate Cleanup)
 
-### Debt Item: Linting Debt (Ruff/UP)
-**Category:** Quality
-**Impact:** Low
-**Effort:** M
-**Resolution plan:** Systematic application of `ruff --fix` to resolve hundreds of `UP` (Upgrade), `I` (Import), and `F` (Pyflakes) errors.
-**Owner:** Jules05 (Immediate cleanup)
-
-### Debt Item: Raw Print Statements
-**Category:** Quality
-**Impact:** Low
+### Debt Item: Inconsistent Risk API in Main Loop
+**Category:** Naming | Quality
+**Impact:** High
 **Effort:** S
-**Resolution plan:** Replace `print()` with `structlog` or `rich.console` across core modules (main.py, explainability, etc.).
-**Owner:** Jules05 (Immediate cleanup)
+**Resolution plan:** Update `main.py` to use the harmonized `validate_signal()` method which returns a structured `RiskDecision` object, replacing the legacy boolean `approve()` method.
+**Owner:** Jules05 (Immediate Cleanup)
 
-### Debt Item: Fragmented Signal Mapping
+### Debt Item: Redundant Position Sizing Logic
 **Category:** Duplication
 **Impact:** Medium
 **Effort:** S
-**Resolution plan:** Replace manual string/integer to SignalDirection mappings with `ModelAction(idx).to_direction()` or `SignalDirection` constants.
-**Owner:** Jules05 (Immediate cleanup)
+**Resolution plan:** Consolidate Kelly Criterion sizing and ATR-based sizing into a single, robust `calculate_position_size` method within the harmonized `RiskManager`.
+**Owner:** Jules05 (Immediate Cleanup)
 
-### Debt Item: Placeholder Secrets in Validator
-**Category:** Quality
-**Impact:** Medium
+### Debt Item: Stale Imports and Exports
+**Category:** Dead Code
+**Impact:** Low
 **Effort:** S
-**Resolution plan:** Ensure `ConfigValidator` strictly rejects placeholder passwords ("password", "change_me") in production-like environments.
-**Owner:** Jules03 (Governance)
+**Resolution plan:** Cleanup `__init__.py` files and remove redundant exports (e.g., `RiskEngine`, misplaced `FeatureEngineer`).
+**Owner:** Jules05 (Immediate Cleanup)
 
-### Debt Item: Fragmented LSTM Architecture
-**Category:** Fragmentation
-**Impact:** High
-**Effort:** M
-**Resolution plan:** Relocate `LSTMAttentionModel` from `src/models/ensemble.py` to `src/models/lstm_model.py` to centralize sequence modeling logic.
-**Owner:** Jules05 (Immediate cleanup)
-
-### Debt Item: Redundant Signal/Action Mapping
-**Category:** Duplication
-**Impact:** Medium
-**Effort:** S
-**Resolution plan:** Replace manual `IntEnum` to index mappings with centralized `ModelAction(idx).to_direction()` calls across the codebase.
-**Owner:** Jules05 (Immediate cleanup)
-
-### Debt Item: Placeholder RL Agents (Dreamer/PPO stubs)
-**Category:** Fragmentation
-**Impact:** Medium
-**Effort:** L
-**Resolution plan:** Schedule for Jules04 (Quant Research) to replace placeholders with functional world-model implementations.
-**Owner:** Jules04
-
-### Debt Item: Inconsistent Model Output Permutations
-**Category:** Fragmentation
-**Impact:** High
-**Effort:** M
-**Resolution plan:** Standardize all model outputs to `[HOLD, BUY, SELL]` and remove legacy permutation logic in `EnsembleModel`.
-**Owner:** Jules05 (Immediate cleanup)
-
-### Debt Item: Placeholder Reward Logic in Trading Environment
-**Category:** Quality
-**Impact:** Medium
-**Effort:** M
-**Resolution plan:** Schedule for Jules01/Jules04 to implement proper reward shaping based on risk-adjusted returns.
-**Owner:** Jules01/Jules04
+---
+*Last Updated: 2026-05-17*
