+++ tests/test_system_bootstrap_to_execution.py
@@ -21,7 +21,7 @@ from sqlalchemy import select
 from src.core.audit_log import AuditEntry, AuditLogger
 from src.core.config import get_config
 from src.core.constants import SignalDirection
-from src.core.feature_engineering import FeatureEngineer
+from src.data.feature_engineering import FeatureEngineer
 from src.core.health import ComponentStatus, HealthStatus, init_health_checker
 from src.core.schemas import TradeSignal
 from src.core.trade_logger import RiskEvent, Trade, TradeLogger
